Initial functionality updates

+++ package360/src/ReadFile.java
@@ -8,23 +8,23 @@ import java.nio.file.*;
 import java.io.*;
 
 class ReadFile {
-    static List<Integer> readFileByName(String fileName) {
+    static List<Float> readFileByName(String fileName) {
 
         List<String> lines;
         String[] elements = null;
-        List<Integer> result = new ArrayList<>(Collections.emptyList());
+        List<Float> result = new ArrayList<>(Collections.emptyList());
         try {
             lines = Files.readAllLines(Paths.get(fileName), StandardCharsets.UTF_8);
 
             for (String s : lines) {
                 if (!s.isBlank())
                     if (!s.contains(",")) {
-                        result.add(Integer.parseInt(s));
+                        result.add(Float.parseFloat(s));
                     } else {
                         elements = s.split(",");
                         for (String i : elements) {
                             if (!i.isBlank())
-                                result.add(Integer.parseInt(i));
+                                result.add(Float.parseFloat(i));
                         }
                         elements = null;
                     }
